requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-appconf==1.1.0
django-compressor==4.5.1
django-countries==7.6.1
django-environ==0.11.2
django-summernote==0.8.20.0
django-tailwind==4.0.1
idna==3.10
joblib==1.4.2
jwt==1.3.1
lxml==5.4.0
nltk==3.8.1
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyPDF2==3.0.0
python-docx==0.8.11
python-dotenv==1.1.0
rcssmin==1.1.2
regex==2024.11.6
requests==2.32.3
rjsmin==1.2.2
sentry-sdk==2.27.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
